For Fabric, send message to get Authorization token from iframe parent
