+++ src/hooks/useKnockoutExplorer.ts
@@ -10,7 +10,7 @@ import { useEffect, useState } from "react";
 import { AuthType } from "../AuthType";
 import { AccountKind, Flights } from "../Common/Constants";
 import { normalizeArmEndpoint } from "../Common/EnvironmentUtility";
-import { sendMessage, sendReadyMessage } from "../Common/MessageHandler";
+import { handleCachedDataMessage, sendMessage, sendReadyMessage } from "../Common/MessageHandler";
 import { Platform, configContext, updateConfigContext } from "../ConfigContext";
 import { ActionType, DataExplorerAction, TabKind } from "../Contracts/ActionContracts";
 import { MessageTypes } from "../Contracts/ExplorerContracts";
@@ -164,6 +164,10 @@ async function configureFabric(): Promise<Explorer> {
               }
             }
 
+            break;
+          }
+          case "authorizationToken": {
+            handleCachedDataMessage(data);
             break;
           }
           default:
